Update product to be able to display different SKUs
Fix too many requests from this IP issue in Admin pages.

Revert "Update product to be able to display different SKUs"

This reverts commit 64d3e31ea53507492f79c679dd220dacc542a56d.

Update product to be able to display different SKUs

Make the first value of product specification as the default specification when add product to shopping cart in main page.

Update product to be able to display different SKUs

Fix order image display issue with the new product image uploading

Update product to be able to display different SKUs

Fix errors

Update product to be able to display different SKUs

fix issue

Update product to be able to display different SKUs

Update product to be able to display different SKUs

Fix Can not add second specification of the same product to cart.

Update product to be able to display different SKUs

Got 500 error when adding product to cart.

Update product to be able to display different SKUs

Fix the issue of adding incorrect number when add different Specification of the same product.
Currently, the number is cumulative. But the product number of each Specification combination should be caculated indenpentently.

Update product to be able to display different SKUs

The image order should keep the same as  the value order of required Specification.

Update product to be able to display different SKUs

1. Make the main page load the image from the first value of required specification of product.
2. Apply the changes to all other products.

Update product to be able to display different SKUs

1.  Make the product load images from the values of the required Specification.
2. It can display all the images uploaded in each value of the required Specification in Product Quick View.  Click different images will change the selection of the first (required) Specification. And also select the different value of the first (required) Specification will change the image which displays in the main display window.
3. Display different item cards in Shopping Cart, if the product Specification is not totally the same.
4. Remove the Specifications in the bottom of the product quick view since it is already moved up.